src/json/personal_tekezo.json.js -> public/js/personal_tekezo.js
