+++ public/extra_descriptions/personal_tekezo.html
@@ -1,0 +1,70 @@
+<h4>General</h4>
+
+<ul>
+  <li>Launch 1Password when <kbd>fn</kbd> is pressed alone</li>
+  <li>
+    Change <kbd>left_control</kbd> to lazy <kbd>left_control</kbd> (Post <kbd>spacebar</kbd> when pressed alone)
+    <ul>
+      <li>Post <kbd>command + `</kbd> when <kbd>left_control</kbd> is tapped while holding <kbd>left_shift</kbd></li>
+      <li>Post <kbd>command + `</kbd> when <kbd>left_control</kbd> is tapped while holding <kbd>left_command</kbd></li>
+    </ul>
+  </li>
+  <li>Post <kbd>spacebar</kbd> when <kbd>left_shift</kbd> is pressed alone</li>
+  <li>Change input sources with <kbd>left_option + left_command</kbd> and <kbd>left_command + left_option</kbd></li>
+  <li>Switch tabs with <kbd>command + control + f/j</kbd></li>
+  <li>Navigate back/forward with <kbd>command + control + g/h</kbd></li>
+  <li>Capture a selected area with <kbd>f13</kbd></li>
+  <li>Change <kbd>control + i/h</kbd> to <kbd>tab/delete</kbd></li>
+  <li>Change <kbd>left_option + a/e</kbd> to <kbd>home/end</kbd></li>
+  <li>Hold <kbd>-/=</kbd> to type 20 hyphens/equal signs</li>
+</ul>
+
+<h4>Mouse</h4>
+
+<ul>
+  <li>Hold <kbd>button5</kbd> and move the mouse to scroll (Post <kbd>command + click</kbd> when pressed alone)</li>
+</ul>
+
+<h4>App-specific settings</h4>
+
+<h5>Finder</h5>
+
+<ul>
+  <li>Disable <kbd>command + l</kbd></li>
+  <li>Change <kbd>control + q</kbd> to <kbd>command + up arrow</kbd> (Move to the parent folder)</li>
+</ul>
+
+<h5>Terminal</h5>
+
+<ul>
+  <li>Disable <kbd>command + d</kbd>, <kbd>command + f</kbd>, and <kbd>command + o</kbd></li>
+</ul>
+
+<h5>Firefox, Google Chrome, and Safari</h5>
+
+<ul>
+  <li>Disable <kbd>command + d</kbd>, <kbd>command + shift + d</kbd>, and <kbd>command + shift + i</kbd></li>
+  <li>Change <kbd>control + a/e</kbd> to <kbd>command + left/right arrow</kbd></li>
+</ul>
+
+<h5>Visual Studio Code</h5>
+
+<ul>
+  <li>Change <kbd>option + n/p</kbd> to <kbd>control + page down/up</kbd></li>
+  <li>Disable <kbd>control + t</kbd></li>
+</ul>
+
+<h5>Microsoft Excel</h5>
+
+<ul>
+  <li>Change <kbd>command + e</kbd> to <kbd>F2</kbd> (Edit the current cell)</li>
+</ul>
+
+<h5>Microsoft Remote Desktop</h5>
+
+<ul>
+  <li>Change <kbd>left_control + h/d</kbd> to <kbd>delete/forward delete</kbd></li>
+  <li>Change <kbd>left_control + i</kbd> to <kbd>tab</kbd></li>
+  <li>Change <kbd>left_control + b/f/n/p</kbd> to <kbd>left/right/down/up arrow</kbd></li>
+  <li>Change <kbd>left_control + a/e</kbd> to <kbd>home/end</kbd></li>
+</ul>
