--- public/extra_descriptions/personal_tekezo.json.html
@@ -1,38 +1,0 @@
-<ul>
-  <li>Change <kbd>left_control</kbd> to <kbd>lazy left_control</kbd> (Post <kbd>spacebar</kbd> when pressed alone)</li>
-  <ul>
-    <li>Change <kbd>left_shift + left_control</kbd> to <kbd>left_command + `</kbd> when <kbd>left_control</kbd> is tapped</li>
-    <li>Change <kbd>left_command + left_control</kbd> to <kbd>left_command + `</kbd> when <kbd>left_control</kbd> is tapped</li>
-  </ul>
-  <li>Post <kbd>spacebar</kbd> when <kbd>left_shift</kbd> is pressed alone</li>
-  <li>Change input sources by <kbd>left_option + left_command</kbd> and <kbd>left_command + left_option</kbd></li>
-  <li>Launch 1Password when <kbd>fn</kbd> is pressed alone</li>
-</ul>
-
-<ul>
-  <li>Change <kbd>command + button2</kbd> to <kbd>command + tab</kbd></li>
-</ul>
-
-<ul>
-  <li>Change <kbd>command + control + f</kbd> to <kbd>command + shift + [</kbd></li>
-  <li>Change <kbd>command + control + j</kbd> to <kbd>command + shift + ]</kbd></li>
-  <li>Change <kbd>command + control + g</kbd> to <kbd>command + [</kbd></li>
-  <li>Change <kbd>command + control + h</kbd> to <kbd>command + ]</kbd></li>
-</ul>
-
-<ul>
-  <li>Change <kbd>; + j</kbd> to <kbd>mouse button #1</kbd></li>
-  <li>Change <kbd>; + k</kbd> to <kbd>mouse button #3</kbd></li>
-  <li>Change <kbd>; + l</kbd> to <kbd>mouse button #2</kbd></li>
-  <li>Change <kbd>; + i</kbd> to warp mouse pointer to the center of screen</li>
-</ul>
-
-<ul>
-  <li>Change <kbd>button5</kbd> + mouse motion to scroll wheel (Post <kbd>command+click</kbd> if <kbd>button5</kbd> is clicked alone)</li>
-  <li>Change <kbd>button6</kbd> to <kbd>command+click</kbd></li>
-</ul>
-
-<ul>
-  <li>Change <kbd>option + -</kbd> to ----------</li>
-  <li>Change <kbd>option + =</kbd> to ==========</li>
-</ul>
