+++ public/groups.json
@@ -1864,6 +1864,10 @@
       "id": "personal-settings",
       "files": [
         {
+          "path": "js/personal_tekezo.js",
+          "extra_description_path": "extra_descriptions/personal_tekezo.html"
+        },
+        {
           "path": "js/personal_tekezo_launcher_mode_v4.js"
         },
         {
@@ -1979,10 +1983,6 @@
         },
         {
           "path": "json/personal_sunyour.json"
-        },
-        {
-          "path": "json/personal_tekezo.json",
-          "extra_description_path": "extra_descriptions/personal_tekezo.json.html"
         },
         {
           "path": "json/personal_thooams.json"
